Coding style changes
Tried to standardize the coding style to increase readability and
ease-of-use.

Changed opts cell array to use varargin for more MATLAB-standard call
structure.
Applied a uniform spacing scheme to code for easy readability.
Changed ctranspose operators when creating row vectors to transpose.
Saved current axes at start of code to ensure it does not change
between the start and end of running.